requirements.txt
allure-pytest==2.8.0
allure-python-commons==2.8.0
pytest==6.2.2
pytest-html==3.1.1
pytest-instafail==0.4.2
pytest-metadata==1.11.0
pytest-ordering==0.6
pytest-parallel==0.1.0
pytest-reportlog==0.1.2
pytest-dependency==0.5.1
pytest-rerunfailures==10.0
pytest-sugar==0.9.1
pytest-timeout==1.4.2
urllib3==1.26.6
requests==2.25.1
xlrd==2.0.1
lxml==4.6.3
XlsxWriter==1.3.7
xlutils==2.0.0
xlwt==1.3.0
crypto==1.2.0
pandas==1.2.3
configparser==5.0.1
locust==1.1
loguru==0.5.1
PyMySQL==1.0
PyYAML==5.4.1
simplejson==3.17.2
requests-toolbelt==0.9.1
typing-extensions==3.10.0.0
pyrsistent==0.18.0
pbr==5.6.0
Flask==1.0.2
Jinja2==2.10
filetype==1.0.7
faker==8.11.0